complete all 4 chart.

+++ ca1_assignment/analysis1/analysis1.py
@@ -4,9 +4,6 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 
-# Handle load selected column from raw data
-
-
 def region_mapper(town):
     mapper = {
         "SEMBAWANG": "NORTH",
@@ -47,7 +44,8 @@ def calculate_price_per_square_feet(resale_price, square_meter_area):
     return per_sqft
 
 
-os.remove("data/clean_data.csv")
+if os.path.exists('data/clean_data.csv'):
+    os.remove("data/clean_data.csv")
 csv_data1 = np.loadtxt("data/resale-flat-prices-based-on-approval-date-1990-1999.csv", skiprows=1, delimiter=",", dtype=str)
 csv_data2 = np.loadtxt("data/resale-flat-prices-based-on-approval-date-2000-feb-2012.csv", skiprows=1, delimiter=",", dtype=str)
 csv_data3 = np.loadtxt("data/resale-flat-prices-based-on-registration-date-from-mar-2012-to-dec-2014.csv", skiprows=1, delimiter=",", dtype=str)
@@ -59,7 +57,7 @@ print("Data Count (Before Concatenate): " + str(combined_data_count))
 
 combined_data = np.concatenate((csv_data1, csv_data2, csv_data3, csv_data4, csv_data5), axis=0)
 del csv_data1, csv_data2, csv_data3, csv_data4, csv_data5
-print("Original Data counts: {}".format(len(combined_data)))
+print("Data Count (Before Concatenate): {}".format(len(combined_data)))
 
 # Create column Region
 region = []
@@ -87,20 +85,13 @@ saved_data = np.genfromtxt("data/clean_data.csv",
                            )
 print("Saved Data counts: {}".format(len(saved_data)))
 
-# ======================
-# DATASET: 1
-# GRAPH: 1 (LINE CHART)
-# ======================
 chart1_data = np.copy(saved_data)
-
-# Transform yyyy-mm to yyyy
 chart1_data['purchase_date'] = [datetime.strptime(date, '%Y-%m').year for date in chart1_data['purchase_date']]
 
 years = np.unique(chart1_data['purchase_date'])
 regions = np.unique(chart1_data['region'])
 
 data_price_by_region = []
-
 for region in regions:
     data_by_region = []
     for year in years:
@@ -132,49 +123,4 @@ for i in data_price_by_region:
     plt.plot(i, label=regions[count], color=color[count])
     count += 1
 legend = plt.legend(loc='upper left', shadow=True)
-plt.show()
-del chart1_data
-
-# ======================
-# DATASET: 1
-# GRAPH: 2 (BOXPLOT)
-# ======================
-town_list = np.unique(saved_data['town'])
-town_count = 1
-for town in town_list:
-    print("{}. {}".format(str(town_count), town), end="\n")
-    town_count += 1
-
-town_choice = int(input("Please select your preferred town (key in numbering): "))
-town_index = town_choice - 1
-town_selected = town_list[town_index]
-if town_choice > len(town_list) or town_choice < 1:
-    print("Sorry, you have entered an invalid choice")
-    print("Unable to continue. Exiting program....")
-
-filtered_data_by_flat_type = saved_data[np.isin(saved_data['flat_type'], ['2 ROOM', '3 ROOM', '4 ROOM', '5 ROOM', 'EXECUTIVE'])]
-hdb_type = np.unique(filtered_data_by_flat_type['flat_type'])
-
-x_label = []
-data_by_room_type = []
-
-for type in hdb_type:
-    filtered_data = filtered_data_by_flat_type[np.isin(filtered_data_by_flat_type['flat_type'], [type]) &
-                                               np.isin(filtered_data_by_flat_type['town'], [town_selected])]
-
-    has_elements = filtered_data.size > 0
-    if has_elements:
-        x_label.append(type)
-        resale_price = filtered_data['resale_price']
-        square_meter_area = filtered_data['floor_area_sqm']
-        per_sqft = calculate_price_per_square_feet(resale_price, square_meter_area)
-        data_by_room_type.append(per_sqft)
-
-# Graph 2: Boxplot (Cosmetic)
-plt.suptitle('HDB RESALE PRICE in {}'.format(town_selected), fontsize=14, fontweight='bold')
-plt.title('Average Resale Price per Square Feet (sqft) by Room Type')
-plt.xlabel('Room Type')
-plt.ylabel('Price (per sqft)')
-
-plt.boxplot(data_by_room_type, labels=x_label)
 plt.show()
